// src/gencode/protos-frontend/enums/onboarding_stages.ts
// @generated by protobuf-ts 2.9.4
// @generated from protobuf file "protos-frontend/enums/onboarding_stages.proto" (package "enums", syntax proto3)
// tslint:disable
/**
 * @generated from protobuf enum enums.OnboardingStage
 */
export enum OnboardingStage {
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_UNIDENTIFIED = 0;
   */
  ONBOARDING_STAGE_UNIDENTIFIED = 0,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_EMAIL_VERIFICATION = 1;
   */
  ONBOARDING_STAGE_EMAIL_VERIFICATION = 1,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_PHONE_VERIFICATION = 2;
   */
  ONBOARDING_STAGE_PHONE_VERIFICATION = 2,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_PAN_VERIFICATION = 3;
   */
  ONBOARDING_STAGE_PAN_VERIFICATION = 3,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_AADHAR_VERIFICATION = 4;
   */
  ONBOARDING_STAGE_AADHAR_VERIFICATION = 4,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_LIVENESS = 5;
   */
  ONBOARDING_STAGE_LIVENESS = 5,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_EXTRA_DETAILS = 6;
   */
  ONBOARDING_STAGE_EXTRA_DETAILS = 6,
  /**
   * @generated from protobuf enum value: ONBOARDING_STAGE_HOME_STRETCH = 7;
   */
  ONBOARDING_STAGE_HOME_STRETCH = 7,
  /**
   * @generated from protobuf enum value: ONBOARDING_COMPLETED = 8;
   */
  ONBOARDING_COMPLETED = 8,
}